+++ scripts/index.ts
@@ -1,7 +1,12 @@
-// Code based on https://github.com/biomejs/biome/discussions/3#discussioncomment-7910787. Thanks, Dani Guardiola!
+// Code initially based on the html tables extraction: https://github.com/biomejs/biome/discussions/3#discussioncomment-7910787.
+// Thanks, Dani Guardiola!
 
 import fs from "fs"
+import path from "path"
 import { JSDOM } from "jsdom"
+
+const extraRulesToDisable = ["simple-import-sort/imports"]
+const rootPath = path.resolve(import.meta.dir, "..")
 
 const getTdString = (row: Element, column: number) =>
   (
@@ -10,9 +15,7 @@ const getTdString = (row: Element, column: number) =>
 
 type Plugin = { id: string; prefix: string }
 
-/**
- * Returns the ESLint rules for the equivalent Biome's rules that are recommended (proof of concept for now)
- */
+/** Returns the ESLint rules for the equivalent Biome's rules that are recommended */
 const getEslintEquivalentRulesForPlugin = (
   plugin: Plugin,
   document: Document,
@@ -62,6 +65,27 @@ const getEslintEquivalentRules = async (): Promise<Array<string>> => {
   )
 }
 
+/** https://typescript-eslint.io/rules/#extension-rules */
+const getTsExtensionsForRules = (rules: Array<string>): Array<string> => {
+  const glob = new Bun.Glob("*.js")
+  const tsExtensionRules = [
+    ...glob.scanSync(
+      path.resolve(
+        rootPath,
+        "node_modules/@typescript-eslint/eslint-plugin/dist/rules",
+      ),
+    ),
+  ].map((s) => s.replace(".js", ""))
+
+  const rulesToExtend = tsExtensionRules.filter((tsExtensionRule) =>
+    rules.includes(tsExtensionRule),
+  )
+
+  return rulesToExtend.map(
+    (ruleToExtend) => `@typescript-eslint/${ruleToExtend}`,
+  )
+}
+
 const writeFile = (rules: Array<string>) => {
   const text = `/**
  * File automatically created by scripts/index.ts.
@@ -78,12 +102,12 @@ ${rules.map((rule) => `    "${rule}": "off",`).join("\n")}
   fs.writeFileSync("index.js", text)
 }
 
-const extraRulesToDisable = ["simple-import-sort/imports"]
-
 const main = async () => {
   const rules = [...(await getEslintEquivalentRules()), ...extraRulesToDisable]
+  const rulesWithTsExtends = [...rules, ...getTsExtensionsForRules(rules)]
+  const rulesNoDuplicates = [...new Set(rulesWithTsExtends)]
 
-  writeFile(rules)
+  writeFile(rulesNoDuplicates)
 
   console.log("Generated index.js!")
 }